--- src/frontend/containers/AdvancedSearch/search.scss
@@ -1,51 +1,0 @@
-#search-form {
-  .criteria {
-    padding: 2px;
-    margin-bottom: 0.25rem;
-    display: flex;
-    align-items: center;
-
-    > :not(.btn) {
-      flex-grow: 1;
-      margin-right: 0.5rem;
-      max-width: 25%;
-    }
-
-    > :nth-child(3) {
-      max-width: none;
-      margin-right: 0.25rem;
-    }
-
-    > .btn {
-      flex-grow: 0;
-      height: 1.5rem;
-      width: 1.5rem;
-    }
-  }
-}
-
-#functions-bar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  [role='radiogroup'] {
-    > div {
-      flex-direction: row;
-      align-items: center;
-    }
-
-    legend {
-      margin: 0;
-      margin-right: 1.5rem;
-    }
-
-    label {
-      margin: 0;
-
-      &:first-of-type {
-        margin-right: 1rem;
-      }
-    }
-  }
-}
